Split out extensions from common.c/.h
