+++ src/support/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(morpho
     PRIVATE
         common.c     common.h
+        extensions.c extensions.h
         memory.c     memory.h
         parse.c      parse.h
         random.c     random.h
